add: app of service

+++ service/app.go
@@ -3,7 +3,9 @@ package service
 import (
 	"bufio"
 	"encoding/json"
+	"errors"
 	"fmt"
+	srv "github.com/kardianos/service"
 	"github.com/mitchellh/go-ps"
 	"github.com/rs/xid"
 	"github.com/sirupsen/logrus"
@@ -18,15 +20,28 @@ import (
 )
 
 var DefaultAppManager *AppManager
+var AppLogger = logrus.New().WithField("scope", "AppManager")
+var NotFound = errors.New("service is nil")
 
 const (
 	StatusStop = iota + 1
 	StatusRunning
 )
 
+const (
+	AppTypeRunnable = "Runnable"
+	AppTypeService  = "Service"
+)
+
 var StatusTextMapping map[int]string = map[int]string{
 	StatusStop:    "Stop",
 	StatusRunning: "Running",
+}
+
+var ServiceStatusMapping = map[srv.Status]int{
+	srv.StatusRunning: StatusRunning,
+	srv.StatusStopped: StatusStop,
+	srv.StatusUnknown: StatusStop,
 }
 
 type AppManager struct {
@@ -45,13 +60,21 @@ func (m *AppManager) GetAppByIdApp(id string) *App {
 func (m *AppManager) RunApp(id string) error {
 	app := m.GetAppByIdApp(id)
 	if app != nil {
-		cmd, err := app.Run()
-		if err != nil {
-			return err
-		}
 		m.Lock()
-		app.Cmd = cmd
-		m.Unlock()
+		defer m.Unlock()
+		if app.Type == AppTypeRunnable {
+			cmd, err := app.RunCommand()
+			if err != nil {
+				return err
+			}
+			app.Cmd = cmd
+		}
+		if app.Type == AppTypeService {
+			err := app.RunService()
+			if err != nil {
+				return err
+			}
+		}
 	}
 	return nil
 }
@@ -84,21 +107,12 @@ func (m *AppManager) StopApp(id string) error {
 
 func (m *AppManager) RunProcessKeeper() {
 	go func() {
-		logrus.Info("running process keeper")
+		AppLogger.Info("running process keeper")
 		for {
 			<-time.After(1 * time.Second)
 			m.Lock()
 			for _, app := range m.Apps {
-				if app.Cmd != nil {
-					process, err := ps.FindProcess(app.Cmd.Process.Pid)
-					if err != nil || process == nil {
-						app.Status = StatusStop
-					}
-					app.Status = StatusRunning
-				} else {
-					app.Status = StatusStop
-				}
-
+				app.UpdateStatus()
 			}
 			m.Unlock()
 		}
@@ -106,16 +120,30 @@ func (m *AppManager) RunProcessKeeper() {
 }
 
 type App struct {
-	Id           string    `json:"-"`
-	AppName      string    `json:"app_name"`
-	StartCommand string    `json:"start_command"`
-	AutoStart    bool      `json:"auto_start"`
-	Dir          string    `json:"-"`
-	Status       int       `json:"-"`
-	Cmd          *exec.Cmd `json:"-"`
+	Id           string      `json:"-"`
+	Type         string      `json:"type"`
+	ServiceName  string      `json:"service_name"`
+	AppName      string      `json:"app_name"`
+	StartCommand string      `json:"start_command"`
+	AutoStart    bool        `json:"auto_start"`
+	Dir          string      `json:"-"`
+	Status       int         `json:"-"`
+	Cmd          *exec.Cmd   `json:"-"`
+	Service      srv.Service `json:"-"`
 }
 
-func (a *App) Run() (*exec.Cmd, error) {
+func (a *App) RunService() error {
+	if a.Service == nil {
+		return NotFound
+	}
+	appService, _ := GetServiceByName(strings.ReplaceAll(a.ServiceName, ".service", ""))
+	return appService.Start()
+}
+func (a *App) StopService() error {
+	appService, _ := GetServiceByName(strings.ReplaceAll(a.ServiceName, ".service", ""))
+	return appService.Stop()
+}
+func (a *App) RunCommand() (*exec.Cmd, error) {
 	parts := strings.Split(a.StartCommand, " ")
 	arg := make([]string, 0)
 	if len(parts) > 1 {
@@ -131,12 +159,40 @@ func (a *App) Run() (*exec.Cmd, error) {
 	return cmd, nil
 }
 func (a *App) Stop() error {
-	if a.Cmd != nil {
+	if a.Cmd != nil && a.Type == AppTypeRunnable {
 		err := a.Cmd.Process.Kill()
 		if err != nil {
 			return err
 		}
 	}
+	if a.Service != nil && a.Type == AppTypeService {
+		err := a.StopService()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (a *App) UpdateStatus() error {
+	if a.Type == AppTypeRunnable && a.Cmd != nil {
+		process, err := ps.FindProcess(a.Cmd.Process.Pid)
+		if err != nil || process == nil {
+			a.Status = StatusStop
+		}
+		a.Status = StatusRunning
+		return nil
+	}
+	if a.Type == AppTypeService && a.Service != nil {
+		status, err := a.Service.Status()
+		if err != nil {
+			a.Status = StatusStop
+			return err
+		}
+		a.Status = ServiceStatusMapping[status]
+		return nil
+	}
+	a.Status = StatusStop
 	return nil
 }
 
@@ -155,7 +211,12 @@ func (a *App) SetAutoStart(isAutoStart bool) error {
 	err := a.SaveConfig()
 	return err
 }
-
+func GetServiceByName(name string) (target srv.Service, err error) {
+	target, err = srv.New(nil, &srv.Config{
+		Name: name,
+	})
+	return
+}
 func LoadApps() error {
 	apps := make([]*App, 0)
 	file, err := os.Open("apps")
@@ -175,13 +236,47 @@ func LoadApps() error {
 
 		err = utils.ReadJson(configPath, app)
 		if err != nil {
-			logrus.Error(err)
+			AppLogger.Error(err)
 			continue
 		}
-		if app.AutoStart {
-			cmd, _ := app.Run()
-			app.Cmd = cmd
+		switch app.Type {
+		case AppTypeRunnable:
+			if app.AutoStart {
+				cmd, _ := app.RunCommand()
+				app.Cmd = cmd
+			}
+		case AppTypeService:
+			appService, err := GetServiceByName(app.ServiceName)
+			if err != nil {
+				AppLogger.Error(err)
+				continue
+			}
+			app.Service = appService
+			status, err := appService.Status()
+			if err != nil {
+				//no service
+				AppLogger.WithFields(logrus.Fields{
+					"App":         app.AppName,
+					"ServiceName": app.ServiceName,
+					"on":          "Get service status",
+				}).Error(err)
+				continue
+			}
+			if app.AutoStart && status == srv.StatusStopped {
+				err = app.RunService()
+				if err != nil {
+					// auto start error
+					AppLogger.WithFields(logrus.Fields{
+						"App":         app.AppName,
+						"ServiceName": app.ServiceName,
+						"on":          "Autostart app",
+					})
+					logrus.Error(err)
+					continue
+				}
+			}
 		}
+
 		apps = append(apps, app)
 	}
 
@@ -189,11 +284,10 @@ func LoadApps() error {
 		return err
 	}
 
-	logrus.Info(fmt.Sprintf("success load %d apps", len(apps)))
+	AppLogger.Info(fmt.Sprintf("success load %d apps", len(apps)))
 	DefaultAppManager = &AppManager{
 		Apps: apps,
 	}
-
 	DefaultAppManager.RunProcessKeeper()
 	return nil
 }
